LPS-71164 Add interface SourceCheck, so we can add other types of checks later
