+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/BaseFileCheck.java
@@ -21,19 +21,15 @@ import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.source.formatter.BNDSettings;
-import com.liferay.source.formatter.SourceFormatterMessage;
 import com.liferay.source.formatter.checks.comparator.ElementComparator;
 import com.liferay.source.formatter.checks.util.SourceUtil;
 import com.liferay.source.formatter.util.FileUtil;
 
 import java.io.File;
 
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.dom4j.Element;
@@ -43,58 +39,16 @@ import org.dom4j.Text;
 /**
  * @author Hugo Huijser
  */
-public abstract class BaseFileCheck implements FileCheck {
-
-	@Override
-	public Set<SourceFormatterMessage> getSourceFormatterMessage(
-		String fileName) {
-
-		if (_sourceFormatterMessagesMap.containsKey(fileName)) {
-			return _sourceFormatterMessagesMap.get(fileName);
-		}
-
-		return Collections.emptySet();
-	}
+public abstract class BaseFileCheck
+	extends BaseSourceCheck implements FileCheck {
 
 	@Override
 	public String process(String fileName, String absolutePath, String content)
 		throws Exception {
 
-		_sourceFormatterMessagesMap.remove(fileName);
+		clearSourceFormatterMessages(fileName);
 
 		return doProcess(fileName, absolutePath, content);
-	}
-
-	protected void addMessage(String fileName, String message) {
-		addMessage(fileName, message, -1);
-	}
-
-	protected void addMessage(String fileName, String message, int lineCount) {
-		addMessage(fileName, message, null, lineCount);
-	}
-
-	protected void addMessage(
-		String fileName, String message, String markdownFileName) {
-
-		addMessage(fileName, message, markdownFileName, -1);
-	}
-
-	protected void addMessage(
-		String fileName, String message, String markdownFileName,
-		int lineCount) {
-
-		Set<SourceFormatterMessage> sourceFormatterMessages =
-			_sourceFormatterMessagesMap.get(fileName);
-
-		if (sourceFormatterMessages == null) {
-			sourceFormatterMessages = new TreeSet<>();
-		}
-
-		sourceFormatterMessages.add(
-			new SourceFormatterMessage(
-				fileName, message, markdownFileName, lineCount));
-
-		_sourceFormatterMessagesMap.put(fileName, sourceFormatterMessages);
 	}
 
 	protected void checkElementOrder(
@@ -444,7 +398,5 @@ public abstract class BaseFileCheck implements FileCheck {
 
 	private final Map<String, BNDSettings> _bndSettingsMap =
 		new ConcurrentHashMap<>();
-	private final Map<String, Set<SourceFormatterMessage>>
-		_sourceFormatterMessagesMap = new ConcurrentHashMap<>();
 
 }
