+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/BaseSourceCheck.java
@@ -1,0 +1,80 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.source.formatter.checks;
+
+import com.liferay.source.formatter.SourceFormatterMessage;
+
+import java.util.Collections;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.concurrent.ConcurrentHashMap;
+
+/**
+ * @author Hugo Huijser
+ */
+public abstract class BaseSourceCheck implements SourceCheck {
+
+	@Override
+	public Set<SourceFormatterMessage> getSourceFormatterMessage(
+		String fileName) {
+
+		if (_sourceFormatterMessagesMap.containsKey(fileName)) {
+			return _sourceFormatterMessagesMap.get(fileName);
+		}
+
+		return Collections.emptySet();
+	}
+
+	protected void addMessage(String fileName, String message) {
+		addMessage(fileName, message, -1);
+	}
+
+	protected void addMessage(String fileName, String message, int lineCount) {
+		addMessage(fileName, message, null, lineCount);
+	}
+
+	protected void addMessage(
+		String fileName, String message, String markdownFileName) {
+
+		addMessage(fileName, message, markdownFileName, -1);
+	}
+
+	protected void addMessage(
+		String fileName, String message, String markdownFileName,
+		int lineCount) {
+
+		Set<SourceFormatterMessage> sourceFormatterMessages =
+			_sourceFormatterMessagesMap.get(fileName);
+
+		if (sourceFormatterMessages == null) {
+			sourceFormatterMessages = new TreeSet<>();
+		}
+
+		sourceFormatterMessages.add(
+			new SourceFormatterMessage(
+				fileName, message, markdownFileName, lineCount));
+
+		_sourceFormatterMessagesMap.put(fileName, sourceFormatterMessages);
+	}
+
+	protected void clearSourceFormatterMessages(String fileName) {
+		_sourceFormatterMessagesMap.remove(fileName);
+	}
+
+	private final Map<String, Set<SourceFormatterMessage>>
+		_sourceFormatterMessagesMap = new ConcurrentHashMap<>();
+
+}
